+++ src/olympusmn_int.cpp
@@ -1400,8 +1400,8 @@ namespace Exiv2 {
         }
 
         char ch;
-        int size = value.size();
-        for (int i = 0; i < size && ((ch = static_cast<char>(value.toInt64(i))) != '\0'); i++) {
+        size_t size = value.size();
+        for (size_t i = 0; i < size && ((ch = static_cast<char>(value.toInt64(i))) != '\0'); i++) {
             os << ch;
         }
         return os;
